fix: don't cleanup .gitkeep file

+++ src/dolly/main.py
@@ -39,6 +39,9 @@ def clean_up() -> None:
             if item.is_dir():
                 for subitem in item.rglob("*"):
                     if subitem.is_file():
+                        #: skip .gitkeep file
+                        if subitem.name == ".gitkeep":
+                            continue
                         subitem.unlink()
                 item.rmdir()
             elif item.is_file():
